Finished !

+++ EksamensOpgaveOOP/IStregsystem.cs
@@ -4,7 +4,7 @@ using System;
 namespace Stregsystemet {
     public interface IStregsystem {
         IEnumerable<Product> ActiveProducts { get; }
-        InsertCashTransaction AddCreditsToAccount(User user, int amount);
+        InsertCashTransaction AddCreditsToAccount(User user, double amount);
         BuyTransaction BuyProduct(User user, Product product);
         Product GetProductByID(int id);
         IEnumerable<Transaction> GetTransactions(User user, int count);
